reduce docker image size by almost half
This commit substantially reduces the size of the docker image by using
multi-stage builds. This allows us to build the software in the first
stage and copy the built artifacts into the second, final stage. We only
install runtime dependencies into the second stage. We also use a python
base image so that python/pip are already available in the image.